Combination / Allocation associations to security: set to non-lazy / outjoin finder methods: join fetch master, strategy, allocations and security
SteppingLimitOrder: modify setDefaultLimits (use only bid and ask)

RuleService.addOrderCallback: pass strategyName

ConfigurationUtil: strategy config overwrites base config

conf-base: remove obsolete configs

git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@4513 19eef36c-86f0-de11-afab-0004ac968479

+++ code/src/main/java/com/algoTrader/service/ib/IBUtil.java
@@ -37,7 +37,7 @@ public class IBUtil {
             contract.m_secType = "OPT";
             contract.m_exchange = IBMarketConverter.marketToString(stockOption.getSecurityFamily().getMarket());
             contract.m_currency = stockOption.getSecurityFamily().getCurrency().toString();
-            contract.m_strike = stockOption.getStrike().intValue();
+            contract.m_strike = stockOption.getStrike().doubleValue();
             contract.m_right = stockOption.getType().toString();
             contract.m_multiplier = String.valueOf(stockOption.getSecurityFamily().getContractSize());
